Fix wrong apparent type for lists
The apparent type of list was determined to be `Inferred(List)`, but this
type is actually less precise that the type that would be inferred,
which can be `List Num` for example. The typechecker was wrongly giving
less precise type to binding of the form `val = [ .. ]` inside
statically typed code.