Edit README
Edit for conciseness and consistency. Remove `nickel$`, which is
a confusing and rare convention for shell commands (as opposed to just
`$`). Used more assertive language: better to say what we are than say
we will someday be something.

Changed the key features keywords to focus on user-relevant things,
like composability etc.