Fix recursive overriding with common fields
The fixpoint computation of a recursive record was wrong in the case
where we are merging two records with common fields, and the fields are
the same priority (i.e. one is not erasing the other, and the resulting
field will be the merging of the two). We needed to patch the thunks
from the original records, and not the ones of the end result.