Fix interpolation in multiline strings
Fix issue #596. The main problem was a `token` annotation instead of a
`regex` in the lexer. Since this bug was preventing the corresponding
code path in the lexer to not be triggered, this patch also fixes a few
other small issues that appeared along the way. Bump the `logos`
dependency and add a regression test as well.