Hide builtin contracts impl within the environment
The implementation of builtin contracts is pure Nickel code written as
part of the `contract.ncl` module. However it pollutes the user environment
with identifiers like `list`, `record`, etc. This PR prefixes those
contract implementation with a `$`, which makes them inaccessible from user code
directly, as an identifier can't start with `$`, to avoid unfortunate
clashes.