Use free variables to limit memory leaks
This commit adds a new transformation pass that annotates each field of
a recursive records with its dependencies on the other fields. This
information is also stored in the associated revertible thunks, which
makes it possible to only add relevant fields to the environment when
performing the fixpoint computation. This should get rid of Rc cycles
causing leaks for everything but actually mutually recursive values.