Hide enum types
Enum type syntax has been changed temporarily since the unified syntax
change. While we bikeshed on a new syntax, we don't want people to rely
on the temporary syntax. This PR hides enum types from stdlib types, and
document enum tags but not enum types in the manual. That way, they are
still morally used in the stdlib, but a random user should be able to
use it without knowing about the type.