First version of the type system specification
This commits adds a base specification of the type system of Nickel and
of type inference algorithm in the OTT language, together with the
required flake configuration to compile the specification to a latex
document.

The specification refers to the system described in the RFC004 on
typechecking, and not the current type system at the time of writing.