Fix panic evaluating switch in non strict mode
The evaluation of a switch is done in two steps: first a Term::Switch
construct is evaluated. The lazy arguments (cases and default) are
pushed on the stack, and an Op1::Switch is generated for the next tick
of evaluation. The issue was that the evaluation of the Op1 could push a
strictness marker AFTER the other args have been pushed, making them
inacessible from the evaluation of the switch operator, which expects
those args to be at the top of the stack. So we must push the strictness
marker before pushing the arguments of switch.