Fix wildcard types "leaking" out of strict blocks
This commit adds a `Wildcard` variant to `ApparentType`, which gets
picked by `apparent_type` whenever there's a type annotation that
contains at least one wildcard.  This variant carries both the type
annotation and the apparent type of the underlying value, allowing to
choose one or the other depending on whether evaluation is strict or not.