[Opt/Closurization] First conservative implementation.
This makes it so array concatenation doesn't need to call closurize, by
introducing a flag in ArrayAttrs.

We still assume in most of the code that Arrays (especially thoses
generated by the interpreter) are not closurized, though there may be
room for improvement there.

In order to hide the added ArrayAttrs during serialization (and set it
to a sane default in deserialization) we need a custom implementation
like the ones for Num and Record.