Fix LSP panicking on fields with contracts
Since the refactoring of the typechecker with a distinct walk phase, the
typechecker now also walk contracts and type annotations. This caused
the LSP to panic in a lot of cases (basically, as soon as a contract is
attached to a field), due to scoping issues and breaking some invariants
of the LSP's linearizer.

This commit fixes that by creating a new proper scope() function for
walking contracts (scope_meta). It also removes the old code that used
to (imperfectly) handle contracts directly in the linearizer, dating
back to when the typechecker didn't process annotations.