Refactoring of global typing environment
Get rid of the useless separation between the global typing environment
and the current typing environment. It was done for performance reasons,
when cloning environments was expensive: but now, environments support
sharing, instead of having two separate environments, we can just start
for the initial (previously "global") and insert directly into it.

Doing so, rename "global" to "initial", which is now more faithful.