Adds "opt" field to metadata
Prepare the feature of optional fields by adding an `opt` field to the
MetaValue structure (representing metadata + value). This commit doesn't
make use of it, doesn't add the corresponding syntax in the grammar nor
does it patch (as it should) various record functions, etc. It just add
the fields and do the bare minimum of fixes to get the code to compile
and to pass tests. The optional field feature is incomplete and inactive
at this point.