Enable optional fields
This commit actally makes use of optional fields, previously added to
metadata. It adds the `optional` keyword in the parser, and patches all
record operations to ignore optional fields that don't have a
definition.

Because of the way metavalues current work, "optional not having a
definition" is not a trivial property to decide. Thus, we approximate it
using simple heuristics, than we hope will be good enough in practice.