Introduce array slices
This commit introduces array slices, which is a representation of an
array that allow for sharing a backing array and have different views
into it, with a starting index and ending index. This makes operations
like `tail` or taking a sub-array both constant time and constant
memory.

One downside is that we may retain elements in memory longer, but we
don't really care too much about that for a use-case like the Nickel
interpreter.