Simplify and improve the builtin enum contract
The enum contract was implemented in a heavy way, in a form of
continuation passing style. This was:

1. Inefficient: for an enum of type of size 5, we may have to call 5
   times a simple function that is a wrapper around an "if x == `tag",
   and 5 times the the continuation.
2. Unsound: type variable in tail position are wrapped in a polymorphic
   contract, making the catch-all case of a switch unusable in a typed
   setting.
3. Unhelpful: the error message in case of contract violation could be
   very obscure, pointing to the inner code of this generated chain of
   function applications.

This commit, rather than encoding the row - represented as linked list -
as a chain of function applications, builds a dictionary out of the enum
types and generate a corresponding switch. Checking is now constant time
(indexing in a hashmap), and if the enum type has a tail, the contract
doesn't even check the tag at all.