Abstract over TermEnvironment in contract equality check
The code checking for contract equality is used only in the typechecker
for now. We want to use it during evaluation too, but the environment
has a different representation (with thunks). This commit abstract over
specific environment representation to make the contract equality
checking code agnostic to the specific term environment that is used, to
prepare for coming changes.