Initial context for typechecking
Typechecking is fed an initial typing environment containing the symbols
defined by the stdlib. On the other hand, the term environment, used to
check for contract equality, was initialized empty, as a simplification
for this recent addition. This commit applies the same treatment for
both environment, and provides typechecking with an initial context
(typing env + term env). As a consequence, a few lines added in the REPL
also fix an earlier shortcoming of contract equality checking, which
wasn't effective in the REPL for symbols defined using top-level let.