Don't panic when unifying row tails
When unifying the tails of row types, we only handle a subset of all
possible unification errors. Previously row mismatches were not handled,
however in certain cases (such as the included test case here) we end up
trying to unify a row tail with a unification variable which resolves to
a type which doesn't match, which resulted in a panic. This commit
handles that error, preventing the panic.