Don't panic when a row mismatch occurs while unifying row tails (#847)
When unifying the tails of row types, we only handle a subset of all
possible unification errors. Previously row mismatches were not handled,
however in certain cases (such as the included test case here) we end up
trying to unify a row tail with a unification variable which resolves to
a type that doesn't match, resulting in a panic. This change
handles that error, preventing the panic.

Co-authored-by: Yann Hamdaoui <yann.hamdaoui@tweag.io>