Fix bad lexing of enum tags
Enum tag parsing previously disallowed tags whose names overlapped with
language keywords (e.g. `if was rejected), while also allowing tags to
have an arbitrary amount of space between the backtick and the
identifier.

This commit fixes both of these issues by handling raw identifier tags
and string tags in slightly different ways.

- For raw tags, the entire enum tag is now a single RawEnumTag token.
  This means that the identifier part can never be confused with a
  keyword during parsing, while also ensuring that there is no extra
  whitespace between the backtick and identifier.
- For string tags, we lex a StringEnumTagBegin token, which is just a
  backtick followed by a double quote. This puts the lexer into string
  mode, and we parse the remainder of the identifier as a regular static
  string.

Note that this means that multiline string enum tags are now rejected.
Indeed we explicitly return a lexer error when we encounter a backtick
followed by an attempt to open a multiline string, as without handling
this case we create the possiblility for a term like `` `m%"test"%m`` to
successfully parse, but with a result that is almost certainly not what
the user intended.