Fix panic when parsing invalid use of row tails
Previously Nickel would panic when parsing an expression such as:

```
let x | forall r. { k | Num; r } = {} in x
```

This was happening because the construction of the `InvalidUniRecord`
error unwraps the `pos` field of the problematic `UniRecord`, but that
field was never set in the parser.

This commit sets the `pos` of `UniRecord`s during parsing, thus
preventing the panic.