Add `map_fields` and `filter_map_fields` functions to `RecordData`
These provide a consistent API for actions that occur multiple times
within the codebase - namely, iterating through a record and building a
new record using the values of each field, while optionally also
filtering certain fields out.

This allows the fields to be updated without exposing information about
the rest of the `RecordData` struct to consumers, which should make it
easier to change it in the future.