Row types refactoring
This commit split the previous monolithic type `Types` into separate
kinds, types, record row types, and enum row types, to forbid the
representation of ill-formed types that was previously possible.

The change required reworking a lot of type-related functions as well as
changing the API of `types` and `typechecking` in particular, but the
result seems to be more readable and better organized.