Introduce `%record_diff_fields%` primitive
Previously we used a regular Nickel function to diff the fields of the
records inside the `$record` contract. However, as we are moving to an
"internal" representation of sealed polymorphic record tails, a regular
function will no longer work. This is because there's no way on the
Nickel side to build a new record which is guaranteed to have the sealed
tail of another.

This commit introduces a new primitive, `%record_diff_fields%`, which
returns only the (visible) fields of the first argument which are not
also present in the second argument, while preserving whatever tail the
first argument had.