Revthunk: Rc -> Option for cached closure
Revertible thunks used Rc<Closure> for the cached data. But when the
revertible thunk is first created, the cached data aren't meaningful and
can't be used as it is (it has free variables which need to be bound).
This change better reflects this reality by using an Option, and by
panicking in some illegal patterns (using the cached term before the
revertible thunk is properly constructed or trying to build the cached
data multiple times). Previously, we would silently accept such
operations with the risk of getting spurious unbound variable errors
later on.