Fix multi-stage recursive overriding
Some multistep overriding patterns weren't handled correctly broken (see
https://github.com/tweag/nickel/issues/908). Revertible thunks are to be
understood as function, and the previous implementation was trying to
away with actually introducing explicit function and application, by
patching directly the environment of expressions with free variables.
The examples presented in #908 makes it clear that this simplification
doesn't work for all cases: this commit fixes the issue by introducing
explicit functions and applications in the general case of merging two
record fields, instead of relying on environment patching.