Prevent merging records with sealed tails
It's unclear that this is something that would be possible to do without
violating parametricity, and it's also unclear what utility this would
have in user code. As such, we've decided to simply disallow this.

Attempting to merge a record with a sealed tail now raises a blame error
with the label that was provided at the point the tail was sealed.