Prevent mapping over records with sealed tails
As with merging, it's unclear that this can ever be implemented in a
sufficiently general way that also doesn't violate parametricity.

This is a little more awkward than merging, in that there are
otherwise-valid programs that this makes invalid - for example, calling
a function which uses a dictionary contract with a
polymorphically-sealed record. However, until we discover a user need
for behaviour like this, it still seems to make the most sense to
completely disallow this, as it's easier to weaken the restriction later
than it is to strengthen overly-lax rules.