Don't lose sealed tails when applying `$record` contract
The `$record` contract implementation uses a Nickel function,
`field_diff`, to construct the sub-record of a given record containing
only those fields not present in a second record. The implemenation of
this function uses `array.foldl` with an empty record as its
accumulator.

This is fine when first sealing a record, as the argument is guaranteed
not to have an existing tail. However, `$record` is also applied when a
record term "leaves" a polymorphic contract, at which point it CAN have
a tail which MUST be preserved in order to later be unsealed by the
`$forall_tail` contract.

In order to give this tail to the constructed sub-record, this commit
adds a `%record_empty_with_tail%` primitive, which takes a record as an
argument and builds a new "empty" record with the argument's sealed
tail.