Use a custom error variant for illegal tail access
This commit introduces `EvalError::IllegalPolymorphicTailAccess`, which
is essentially a version of `BlameError` with a custom error reason
which is displayed to the user. This allows us to give more information
when a contract is violated as a result of mapping over or merging a
record with a sealed tail.

Since this new error and `BlameError` are handled near-identically, this
commit also extracts common functionality into a `blame_error` module
within the `error` module.