[refactor] Re-implement polymorphic record contracts (#949)
* Use `RecordData` constructors more consistently

This commit updates a few places where `RecordData` was being directly
constructed to instead call one of its constructor functions.

This reduces the number of places which need knowledge of the internal
structure of the `RecordData` struct, which makes it easier to add new
fields.

* Add `sealed_tail` field to `RecordData`

* Don't lose sealed tails when applying `$record` contract

The `$record` contract implementation uses a Nickel function,
`field_diff`, to construct the sub-record of a given record containing
only those fields not present in a second record. The implemenation of
this function uses `array.foldl` with an empty record as its
accumulator.

This is fine when first sealing a record, as the argument is guaranteed
not to have an existing tail. However, `$record` is also applied when a
record term "leaves" a polymorphic contract, at which point it CAN have
a tail which MUST be preserved in order to later be unsealed by the
`$forall_tail` contract.

In order to give this tail to the constructed sub-record, this commit
adds a `%record_empty_with_tail%` primitive, which takes a record as an
argument and builds a new "empty" record with the argument's sealed
tail.

* Add `%record_seal_tail%` op

This commit adds a new primop, `%record_seal_tail%`. This op takes four
arguments:

- a sealing key which must be provided later to unseal the tail,
- a label which is used to assign blame when the tail is accessed
  illegally,
- the record which we will seal the tail into,
- the term that we wish to seal (note that currently this could be any
  term, though in practice it should always be a record).

It returns the record, with the tail term stored inside it, but
completely invisible to user code.

* Add `%record_unseal_tail%` op

This op is the dual to `%record_seal_tail%` in that, given the correct
sealing key, it returns the tail term that was previously sealed away.

The op takes three arguments:

- the aforementioned sealing key,
- a label which will be used to assign blame if something goes wrong
  while unsealing,
- the record whose tail we want to unseal.

* Prevent merging records with sealed tails

It's unclear that this is something that would be possible to do without
violating parametricity, and it's also unclear what utility this would
have in user code. As such, we've decided to simply disallow this.

Attempting to merge a record with a sealed tail now raises a blame error
with the label that was provided at the point the tail was sealed.

* Prevent mapping over records with sealed tails

As with merging, it's unclear that this can ever be implemented in a
sufficiently general way that also doesn't violate parametricity.

This is a little more awkward than merging, in that there are
otherwise-valid programs that this makes invalid - for example, calling
a function which uses a dictionary contract with a
polymorphically-sealed record. However, until we discover a user need
for behaviour like this, it still seems to make the most sense to
completely disallow this, as it's easier to weaken the restriction later
than it is to strengthen overly-lax rules.

* Reimplement `$forall_tail` contract with new primitives

This commit also rewrites the tests for polymorphic record sealing. In
particular, it makes the failure tests slightly more explicit by
documenting why they are expected to fail, which should help with
debugging.

The success case tests have also been updated, as it is now invalid to
use `record.insert` and `record.remove` with a record under a
polymorphic contract (as they both use dictionary contracts, which
attempt to map over the record). However, these can both be replaced by
their primitive alternatives, to which no contracts are applied, so the
same basic test cases still work.

* Use a custom error variant for illegal tail access

This commit introduces `EvalError::IllegalPolymorphicTailAccess`, which
is essentially a version of `BlameError` with a custom error reason
which is displayed to the user. This allows us to give more information
when a contract is violated as a result of mapping over or merging a
record with a sealed tail.

Since this new error and `BlameError` are handled near-identically, this
commit also extracts common functionality into a `blame_error` module
within the `error` module.

* Prefer struct splat notation where possible

* Use dynamic record types for primop record args

* Raise a type error when trying to seal a tail that isn't a record