Change switch to match
This commit changes the keyword of the switch branching construct to a
match, to make room for future more advanced features like pattern
matching.

Also, the match is now curried, meaning it doesn't have to be applied to
an argument and `match {..cases..}` can be used directly as a function.