Decouple string start & end delimiters in parser logic
Since both multiline and symbolic strings will end with `"%`, it will no
longer be true that the start & end delimiters of a string must match. It
will also no longer be the case that we can infer whether or not
indentation must be stripped from the starting delimiter.

This commit separates the opening and closing delimeter into two
different types, and provides a more explicit API for working with
them, which will be resilient to the addition of a new
`StringStartDelimiter` in a later commit.