Use test-generator to simplify pass test file
This commit uses the test-generator library to automatically generate a
set of integration tests from any `.ncl` file in the
`tests/integration/pass` directory.

As a consequence, simply adding a new file to that directory is
sufficient to add a new test case.

This also meant that the non-test files, like the assertion library and
files that are used to test importing, were moved to a '/lib' directory
so test-generator doesn't pick them up.