Use test-generator to simplify examples tests
This commit uses the test-generator library to automatically extract
test cases from Nickel files in subdirectories of the `examples/` dir.

Because not every `.ncl` file in these directories is a passing test
case, it also introduces a simple annotation to the top of each example
source file indicating whether it should pass, raise a blame error, or
be ignored. These annotations could easily be expanded in future to
cover other types of error.

This automates the process of adding new example test cases, as simply
adding a new unannotated example will cause the tests to fail, prompting
the addition of an annotation and thus a new test case.