Return grapheme cluster count from `%str_length%` primitive
Previously `%str_length%` was returning the number of bytes in the Rust
`String` storage, which lead to unexpected results.

This commit uses the `unicode-segmentation` crate to instead count the
number of extended grapheme clusters in the string. This means that even
complex emoji comprised of multiple characters, but visually rendering
as one, get counted as just one character.

One potential downside to this is that this isn't consistent with how
other standard library functions (such as splitting a string) currently
work, however it also wasn't consistent before, and this change at least
moves things in the right direction