Do not panic on type path mismatch
Before, the type path information encoded in a label and the original
type/contract annotation had to match, by virtue of the Nickel
implementation. We panicked if not the case.

Since the addition of `contract.apply` and the possibility of using
types as terms, this is not true anymore. This PR relaxes the invariant
expected by the error reporting code to avoid panicking in some case
where the user uses `contract.apply` with builtin types.