Fix multiline string interpolation when preceded by a `"` character (#1023)
* Fix multi-line string interpolation rules

Due to an issue with the Logos parser generator, we have a special rule
which looks for a `"` followed by some number of `%`s and then splits
the token into one `"%+` token and an interpolation token.

When symbolic strings were introduced, we changed how the lexer counts
its opening `%`s, to no longer consider the `m` or `s` character as part
of the token. However, we did not update the condition for the rule to
patch over the Logos issue. This essentially meant that we would
introduce an interpolation after a `"` with one too-few opening `%`
characters.

This commit fixes that behaviour by changing the `>=` check to a `>`.

* Re-organise string tests