[REVIEW] Introduce new RuntimeError term
Because of RFC005, missing field definition error are known in advance,
and with better context, which is great. However, in some situation,
they aren't necessary raised in the hand: for example if the field being
treated is not used, we shouldn't raise the error. This commit adds a
new term to the AST, RuntimeError, which is just a delayed error: if it
is never evaluated, it's fine, but if the evaluator ecounters it, the
error it contains is simply raised.