Add --color option (#1033)
This flag has three possible values: auto, always and never. Its default
is 'auto'.

With 'auto' selected, we allow the codespan_reporting and rustyline libs
to determine whether to show colors based on the current terminal
settings.

With 'always' selected, we force-enable colors in both error reporting
and the repl.

With 'never' selected, we disable all colors in output.