[REVIEW] Adapt query to FieldMetadata
Since RFC005, quering just a value doesn't make anymore (random values
don't have metadata). We have to query a field path from a specific
record. This commit reworks the querying infrastructure to handle that.
We recursively evaluate each element of the path to a record, until the
last, from which we extract the corresponding field and access to the
metadata.