Make export snapshot tests capture stderr
Add standard error capture in the snapshot tests for `nickel export`.
This is done by calling `assert_snapshot!` twice, once for `stdout` and
once for `stderr`, so that the snapshots are reasonably reviewable in
`cargo insta review`. Combining both into one call would be feasible,
but only at the cost of presenting the output as raw string literals
during snapshot review.