[REVIEW] 2nd phase of fixing merging
Follow-up of the previus commit to fix merging in the context of RFC005.
In particular, the share_normal_form transformation is udpated to
introduce revertible bindings for field's metadata as well, as the
contract attached to one field may depend recursively on other fields of
the same record. Various missing closurizing/revert/saturate are added
as well.