[REVIEW] Update apply_contract phase for lazy prop
The apply_contract phase doesn't make sense anymore in the lazy
propagation model of RFC005. However, we still need to convert "static"
annotations (the one provided by the user) to pending contracts, the
data that the runtime manipulates and cares about. This commit update
the contract application phase to become a pending contract generation
phase.