Remove redundant check for "indirect" nested record usage.
We already know that `referenced_declaration` refers to the
declaration we want to add a usage to. Any other clever check
to find "indirect" usages becomes unclear and will lead to bugs.
This check for "indirect" usages checks if the referenced field
has a value which is a record, and it then tries to find a field
in this record, which has a name as the actual referenced field,
and it then assumes the nested field is the referenced field.
This is not correct.

For example, in the snippet below `stuff.data` will refer to the
the nested `data` instead of the original `data` field.
```nickel
let stuff = {
    data = {
        data = 1,
        b = 2
    }
} in
stuff.data
```