Don't let declaration term kinds point to an annotated term.
Declaration term kinds should not point to an annotated term
because annotated terms are not linearized; only the contracts
and types contained in the annotations are linearized. If we
let declarations point to annotated term, this will simply mean
that we allow the declaration to point to its contract annotation
instead of its actual value, which, of course, is a bug.