Correctly type identifiers introduced in destructuring patterns (#1099)
* Auto generate passing destructuring tests

* Remove unimplemented `Destruct::Array` variant

This was left unimplemented, which meant that the extra - never parsed
- variant still had to be dealt with elsewhere in the code.

It seems sensible to remove it for now, as it's easy enough to add
back in later on if/when we implement array destructuring.

* Infer types for idents introduced via record destructuring

This commit infers & typechecks identifiers introduced as part of a
destructured record. Previously these idents always had type `Dyn`.

A high-level overview of the process for typing these identifiers looks
like this:

- Infer a type for the destructured record based solely on the pattern.
- Infer a type for the record based on the expression being destructured.
- Unify these two types.
- Iterate through the pattern assigning the unified types to each
  identifier. If the pattern contains a `..rest` pattern, type this
  using the tail of the record, plus any unused "known fields" from
  the destructuring.

Note that this commit does not make use of explicit type annotations
within patterns. These are currently converted to contracts & applied
during evaluation. RFC005, which is currently in-progress, reworks how
such types are represented, so the work to use them in patterns will be
done after that change has landed.

* Ensure nonsensical destructuring is handled gracefully in typechecker