Only parse functions as FunPatterns when they destructure args
Previously any `fun` was parsed as a `FunPattern`, even if they didn't
do any destructuring. This commit changes the parser so that `fun`s
are parsed as `Fun`s unless they actually need to destructure their
arguments.

This means we no longer need to handle `Destruct::Empty` in the
typechecker (and additionally that it can likely be removed entirely
in the near future).

One additional note here is that while working on this commit, we
noticed a bug in the `FunPattern` branch of `walk`, where it was
assigning its argumetn the type of the function body. This had the
side effect of making a test in `typechecking.ncl` pass which should
otherwise have failed. As a result, this commit changes both the `walk`
implementation for `FunPattern`, as well as the test in question.